backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.2
pydantic-settings==2.9.1
python-multipart==0.0.20
google-genai==1.14.0
playwright==1.52.0
beautifulsoup4==4.13.4
python-dotenv==1.1.0
